Move component tests into parent folders
